Move timetables import logic to a separate service to enable more modular and testable code.
